refactor(3): extraer scraper a jobhunter/scraper.py
- scrape_posts: busqueda LinkedIn + scroll + extraccion de emails
- do_linkedin_login (alias _do_linkedin_login): flujo persistente de
  login con espera de 5 min para 2FA
- job.py re-exporta ambos
- tests/test_scraper.py: timeout retorna [], fallback time_filter,
  week/24h en URL, query URL-encoded
